requirements.txt
# MultiTool - All-in-One File Processing Application
# Core requirements for Flask-based multitool with image, video, and PDF processing

# Web Framework
flask==3.0.0
flask-cors==4.0.0

# Image Processing
opencv-python==4.8.1.78
pillow==10.1.0
ultralytics==8.0.206
numpy==1.24.3

# Video Processing  
yt-dlp==2023.10.13
ffmpeg-python==0.2.0
moviepy==1.0.3

# PDF Processing
PyMuPDF==1.23.8
pdfplumber==0.10.3
pypdf==3.0.1

# Utilities
werkzeug==3.0.1
python-magic==0.4.27
requests==2.31.0
python-dotenv==1.0.0

# Optional production dependencies
gunicorn==21.2.0

# Development dependencies (optional)
pytest==7.4.3
pytest-flask==1.3.0
black==23.11.0
flake8==6.1.0
